common/utils/kafka_handler/requirements.txt
kafka-python
docker
redis
psycopg2-binary